pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shroomleur-api"
version = "0.1.0"
description = "API d'authentification et gestion des comptes utilisateurs pour Shroomleur"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "pydantic[email]==2.5.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "google-auth==2.25.2",
    "google-auth-oauthlib==1.1.0",
    "google-auth-httplib2==0.2.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "alembic==1.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.1",
    "ruff==0.1.8",
    "mypy==1.7.1",
]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [".venv", "__pycache__", ".git"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
